+++ lib/widget/aspect_ratio.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/cupertino.dart';
+
+extension ContainerWidgetX on Widget {
+  /// see [AspectRatio]
+  AspectRatio aspectRatio({
+    required double aspectRatio,
+  }) =>
+      AspectRatio(
+        aspectRatio: aspectRatio,
+        child: this,
+      );
+}