+++ lib/widget/hero.dart
@@ -1,11 +1,13 @@
 import 'package:flutter/widgets.dart';
 
 extension HeroWidgetX on Widget {
-  Hero hero(Object tag,
-          {CreateRectTween? createRectTween,
-          HeroFlightShuttleBuilder? flightShuttleBuilder,
-          HeroPlaceholderBuilder? placeholderBuilder,
-          bool transitionOnUserGestures = false}) =>
+  Hero hero(
+    Object tag, {
+    CreateRectTween? createRectTween,
+    HeroFlightShuttleBuilder? flightShuttleBuilder,
+    HeroPlaceholderBuilder? placeholderBuilder,
+    bool transitionOnUserGestures = false,
+  }) =>
       Hero(
           tag: tag,
           createRectTween: createRectTween,
@@ -14,22 +16,21 @@ extension HeroWidgetX on Widget {
           transitionOnUserGestures: transitionOnUserGestures,
           child: this);
 
-  /// If [tag] is null, [this] will be returned.
-  Widget heroOr(Object? tag,
-      {CreateRectTween? createRectTween,
-      HeroFlightShuttleBuilder? flightShuttleBuilder,
-      HeroPlaceholderBuilder? placeholderBuilder,
-      bool transitionOnUserGestures = false}) {
-    if (tag != null) {
-      return Hero(
-          tag: tag,
-          createRectTween: createRectTween,
-          flightShuttleBuilder: flightShuttleBuilder,
-          placeholderBuilder: placeholderBuilder,
-          transitionOnUserGestures: transitionOnUserGestures,
-          child: this);
-    } else {
-      return this;
-    }
-  }
+  /// If [tag] is null, this will be returned.
+  Widget heroOrNot(
+    Object? tag, {
+    CreateRectTween? createRectTween,
+    HeroFlightShuttleBuilder? flightShuttleBuilder,
+    HeroPlaceholderBuilder? placeholderBuilder,
+    bool transitionOnUserGestures = false,
+  }) =>
+      tag == null
+          ? this
+          : Hero(
+              tag: tag,
+              createRectTween: createRectTween,
+              flightShuttleBuilder: flightShuttleBuilder,
+              placeholderBuilder: placeholderBuilder,
+              transitionOnUserGestures: transitionOnUserGestures,
+              child: this);
 }